requirements.txt
flask
boto3
python-decouple